refactor: SearchAndCreatePrescription를 두 개로 분리
